Combine StatementSyntaxComparer and TopSyntaxComparer together (#51021)
